Fixed TranslationConstants issues and renamed file
