+++ src/Moryx.Resources.Web/app/src/app/components/default-details-view/default-details-view.ts
@@ -6,7 +6,7 @@
 import { Component, ChangeDetectionStrategy } from '@angular/core';
 import { EmptyState } from '@moryx/ngx-web-framework/empty-state';
 import { TranslatePipe } from '@ngx-translate/core';
-import { TranslationConstants } from '@app/extensions/translation-constants.extensions';
+import { TranslationConstants } from '@app/extensions/translation-constants';
 
 @Component({
   selector: 'app-default-details-view',
