+++ src/Moryx.Resources.Web/app/src/app/extensions/translation-constants.ts
@@ -52,7 +52,6 @@ export class TranslationConstants {
     REFERENCES_TAB_TITLE: 'DETAILS.REFERENCES_TAB_TITLE',
     METHODS_TAB_TITLE: 'DETAILS.METHODS_TAB_TITLE',
     REFERENCES: {
-      LINK: 'DETAILS.REFERENCES.LINK',
       LINKED_RESOURCE: 'DETAILS.REFERENCES.LINKED_RESOURCE',
       PLACEHOLDER: 'DETAILS.REFERENCES.PLACEHOLDER',
       NO_TARGET: 'DETAILS.REFERENCES.NO_TARGET',
