+++ src/Moryx.Resources.Web/app/src/assets/languages/en.json
@@ -19,7 +19,7 @@
     "CREATE": "CREATE",
     "ADD": "Add resource",
     "ADD_CHILD": "Add child resource",
-    "RELOAD:": "Reload",
+    "RELOAD": "Reload",
     "NEXT": "Next"
   },
   "ADD_RESOURCE_DIALOG": {
@@ -36,7 +36,6 @@
     "REFERENCES_TAB_TITLE": "References",
     "METHODS_TAB_TITLE": "Methods",
     "REFERENCES": {
-      "LINK": "Link",
       "LINKED_RESOURCE": "Linked Resource",
       "PLACEHOLDER": "Select an available resource to add it as a reference",
       "NO_TARGET": "No Target Selected"
